somehow hacky way to stop greacefully

+++ mangalife_downloader.py
@@ -5,6 +5,7 @@ import multiprocessing
 import threading
 import re
 import ast
+import signal
 from zipfile import ZipFile
 import requests
 
@@ -16,12 +17,14 @@ parser.add_argument("urls", nargs="+")
 ARGS = parser.parse_args() # args.picker contains the modality
 
 
-def printer(manga_name: str, printing_queue: multiprocessing.Manager().Queue, number_chapters: int) -> None:
+def printer(manga_name: str, number_chapters: int) -> None:
     """ function that updates the count of the executed chapters """
+    if STOPPER.value:
+        return
     print(f"- {manga_name}: 0 / {number_chapters} completed", end="\r")
     failed = []
     for i in range(1, number_chapters+1):
-        token = printing_queue.get()
+        token = PRINTING_QUEUE.get()
         if token == "quit":
             return
         if token:
@@ -37,9 +40,11 @@ def printer(manga_name: str, printing_queue: multiprocessing.Manager().Queue, nu
         print("No chapter has failed")
 
 
-def download_and_zip(chapter: dict, folder_path: str, printing_queue: multiprocessing.Manager().Queue, manga_name: str) -> bool:
+def download_and_zip(chapter: dict, folder_path: str, manga_name: str) -> bool:
     """ given path and chapter_path, create the zip file
     add a token to the queue when the process is done """
+    if STOPPER.value:
+        return
 
     chapter_name_number = chapter["Chapter"]
     chapter_number = str(int(chapter_name_number[1:-1]))
@@ -58,6 +63,8 @@ def download_and_zip(chapter: dict, folder_path: str, printing_queue: multiproce
         pages = []
         page_number = 0
         while True:
+            if STOPPER.value:
+                return
             page_number += 1
             url_page = f"https://www.manga4life.com/read-online/{manga_name}-chapter-{chapter_number}{index}-page-{page_number}.html"
             response = requests.get(url_page, timeout=10)
@@ -81,23 +88,19 @@ def download_and_zip(chapter: dict, folder_path: str, printing_queue: multiproce
             file_path = os.path.join(chapter_path, f"{page_number:03d}.png")
             if not os.path.exists(file_path):
                 # open the file in binary write mode
-                try:
-                    with open(file_path, "wb") as page:
-                        for chunk in response.iter_content(1024):
-                            page.write(chunk)
-                except KeyboardInterrupt:
-                    if os.path.exists(file_path):
-                        os.remove(file_path)
-                    return
+                with open(file_path, "wb") as page:
+                    for chunk in response.iter_content(1024):
+                        page.write(chunk)
             pages.append(file_path)
 
         # ZIP
-        try:
-            with ZipFile(zip_path, "a") as zip_file:
-                for page in pages:
-                    page_path = os.path.join(chapter_path, page)
-                    zip_file.write(page_path, page)
-        except KeyboardInterrupt:
+        with ZipFile(zip_path, "a") as zip_file:
+            for page in pages:
+                if STOPPER.value:
+                    break
+                page_path = os.path.join(chapter_path, page)
+                zip_file.write(page_path, page)
+        if STOPPER.value:
             if os.path.exists(zip_path):
                 os.remove(zip_path)
             return
@@ -112,60 +115,68 @@ def download_and_zip(chapter: dict, folder_path: str, printing_queue: multiproce
         if not os.path.exists(zip_path):
             failed_number = chapter_name_number
 
-    printing_queue.put(failed_number)
+    PRINTING_QUEUE.put(failed_number)
 
 
-def main() -> None:
+def main(url_manga: str) -> None:
     """ main function """
+    if STOPPER.value:
+        return
 
-    print("Press CTRL+C to quit.")
     # fetch url and chapters data
-    for url in ARGS.urls:
-        manga_name = url.split("/")[-1]
-        response = requests.get(url, timeout=30)
-        if response.status_code != 200:
-            print("cannot request the mange, existing the program")
-            return
-        html_string = response.text
-        chapters_string = re.findall(r"vm.Chapters = (.*);", html_string)[0].replace("null", "None")
-        list_chapters = ast.literal_eval(chapters_string)
+    manga_name = url_manga.split("/")[-1]
+    response = requests.get(url_manga, timeout=30)
+    if response.status_code != 200:
+        print("Cannot request the mange, existing the program")
+        return
+    html_string = response.text
+    chapters_string = re.findall(r"vm.Chapters = (.*);", html_string)[0].replace("null", "None")
+    list_chapters = ast.literal_eval(chapters_string)
 
-        # create folder if does not exists
-        mangas_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "Mangas")
-        os.makedirs(mangas_path, exist_ok=True)
-        folder_path = os.path.join(mangas_path, manga_name)
-        os.makedirs(folder_path, exist_ok=True)
+    # create folder if does not exists
+    mangas_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "Mangas")
+    os.makedirs(mangas_path, exist_ok=True)
+    folder_path = os.path.join(mangas_path, manga_name)
+    os.makedirs(folder_path, exist_ok=True)
 
-        # printing queue for communicating between the printer function and the pool
-        manager = multiprocessing.Manager()
-        printing_queue = manager.Queue()
+    # add more to the list of chapters
+    list_chapters = [[chapter, folder_path, manga_name] for chapter in list_chapters]
+    number_chapters = len(list_chapters)
 
-        # add more to the list of chapters
-        list_chapters = [[chapter, folder_path, printing_queue, manga_name] for chapter in list_chapters]
-        number_chapters = len(list_chapters)
+    # start processing pool
+    pool = multiprocessing.Pool(processes=MAX_PROCESSES)
 
-        # start processing pool
-        pool = multiprocessing.Pool(processes=MAX_PROCESSES)
+    # send all the processes to a pool
+    printer_thread = threading.Thread(target=printer,
+                                        daemon=True,
+                                        args=(manga_name, number_chapters))
+    printer_thread.start()
+    pool.starmap(download_and_zip, list_chapters)
 
-        # send all the processes to a pool
-        printer_thread = threading.Thread(target=printer,
-                                          daemon=True,
-                                          args=(manga_name, printing_queue, number_chapters))
-        printer_thread.start()
-        try: # if CTRL+C stop execution
-            pool.starmap(download_and_zip, list_chapters)
-        except KeyboardInterrupt:
-            print()
-            pool.terminate()
-            print("\nProgram terminated")
-
-        pool.close()
-        pool.join()
-        printing_queue.put("quit")
-        printer_thread.join()
-
-        manager.shutdown()
+    pool.close()
+    pool.join()
+    printer_thread.join()
 
 
-if __name__ == "__main__":
-    main()
+def sigint_handler(sig, frame):
+    """ signal keyboard interrupt handler """
+    STOPPER.value = 1
+    PRINTING_QUEUE.put("quit")
+
+
+# MAIN FUNCTION - keep everything as global for simplicity
+
+STOPPER = multiprocessing.Value("i", 0)
+MANAGER = multiprocessing.Manager()
+PRINTING_QUEUE = MANAGER.Queue()
+
+signal.signal(signal.SIGINT, sigint_handler)
+
+print("Press CTRL+C to quit.")
+for url in ARGS.urls:
+    main(url)
+
+PRINTING_QUEUE.put("quit")
+MANAGER.shutdown()
+
+print("\nProgram terminated." if STOPPER.value else "\nFinished downloading.")
